Fix decimal handling
Fixes errors in decimal handling.
Previously, the display input was converted to
the Number type before being displayed. This meant
that numbers with leading zeros, i.e. 0.003 could
not be entered.
All numbers are now treated as strings until an
operation is applied. This allows for numbers with
leading zeros.