Add overflow limit
Adds a limit to the number of characters the user
can input.

Previously, rounding was used to limit display of
large numbers.
With the introduction of decimal functionality,
numbers are treated as strings until operations
are applied to them. Therefore an overflow limit
is necessary. This is higher than the rounded
output of the operations, to accommodate for the
decimal place.